requirements.txt
fastapi
uvicorn
mysql-connector-python
python-multipart
requests
python-dotenv
httpx